give refresh functionality its own module
